moved classad to expressions and changed access to evaluate
